src/create_scorer.py and src/save_cv_results_plot.py added and model_training.py modified
