+++ scripts/data_download.py
@@ -22,6 +22,7 @@ from src.download_read_combine_data import download_read_combine_data
               help="Do you want to overwrite the file if it exists?", 
               default=False)
 def main(url, save_to, file_to, force_save):
+    
     """Downloads the data from the web to a local filepath and combine it."""
     download_read_combine_data(url, save_to, file_to, force_save)
